# docs/swagger.yaml
basePath: /
definitions:
  models.CrawlRequest:
    properties:
      delay:
        example: 200ms
        type: string
      depth:
        example: 2
        type: integer
      enable_headless:
        example: false
        type: boolean
      enable_html:
        example: true
        type: boolean
      enable_sitemap:
        example: true
        type: boolean
      headless_timeout:
        example: 30
        type: integer
      job_id:
        example: my-custom-session-123
        type: string
      max_urls:
        example: 1000
        type: integer
      url:
        example: https://example.com
        type: string
      wait_for_js:
        example: true
        type: boolean
      workers:
        example: 10
        type: integer
    required:
    - url
    type: object
  models.CrawlResponse:
    properties:
      job_id:
        example: 60f7b3b3b3b3b3b3b3b3b3b3
        type: string
      message:
        example: Crawl job started successfully
        type: string
      status:
        example: accepted
        type: string
    type: object
  models.CrawlResult:
    properties:
      crawled_at:
        type: string
      duration:
        type: string
      id:
        type: string
      settings:
        $ref: '#/definitions/models.CrawlSettings'
      target_url:
        type: string
      total_urls:
        type: integer
      urls:
        items:
          type: string
        type: array
      urls_per_second:
        type: string
    type: object
  models.CrawlSettings:
    properties:
      delay:
        type: string
      depth:
        type: integer
      workers:
        type: integer
    type: object
  models.JobStatus:
    properties:
      created_at:
        example: "2023-07-18T10:30:45Z"
        type: string
      error:
        example: Error message if failed
        type: string
      id:
        example: 60f7b3b3b3b3b3b3b3b3b3b3
        type: string
      progress:
        example: Starting crawl...
        type: string
      request:
        $ref: '#/definitions/models.CrawlRequest'
      result:
        $ref: '#/definitions/models.CrawlResult'
      status:
        example: completed
        type: string
      updated_at:
        example: "2023-07-18T10:32:15Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A fast web crawler API that extracts URLs from websites and stores
    results in MongoDB
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Web Crawler API
  version: "1.0"
paths:
  /crawl:
    post:
      consumes:
      - application/json
      description: Initiates a web crawling job for the specified URL with configurable
        parameters
      parameters:
      - description: Crawl parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CrawlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CrawlResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a new web crawl
      tags:
      - crawl
  /crawls:
    get:
      consumes:
      - application/json
      description: Retrieves a list of recent crawl results from the database
      parameters:
      - default: 10
        description: Maximum number of results to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CrawlResult'
            type: array
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List recent crawl results
      tags:
      - crawls
  /crawls/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific crawl result by its ID
      parameters:
      - description: Crawl ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CrawlResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get specific crawl result
      tags:
      - crawls
  /health:
    get:
      description: Check if the API server is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /jobs:
    get:
      consumes:
      - application/json
      description: Retrieves a list of recent jobs from the database
      parameters:
      - default: 10
        description: Maximum number of results to return
        in: query
        name: limit
        type: integer
      - description: Filter by job status (running, completed, failed)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.JobStatus'
            type: array
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List recent jobs
      tags:
      - jobs
  /jobs/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves the current status and progress of a crawl job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JobStatus'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get crawl job status
      tags:
      - jobs
  /ws/{id}:
    get:
      description: Establishes a WebSocket connection to receive real-time updates
        for a specific crawl job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Connect to live crawl updates
      tags:
      - websocket
swagger: "2.0"